requirements.txt
Django==3.1
django-allauth==0.42.0
django-contrib-comments==1.9.2
django-rest-framework
django-crispy-forms==1.9.2
humanize==2.6.0
Pillow==8.1.2
six==1.15.0
Jinja2==2.11.1
python-decouple==3.3
gunicorn==20.0.4
whitenoise==5.1.0
psycopg2==2.8.5
boto3==1.16.63
django-storages==1.11.1
django-filter==2.4.0
xlwt==1.3.0
python-memcached==1.59
django-imagekit==4.0.2
django-notifications-hq==1.6.0
django-anymail
django-htmlmin==0.11.0
django-honeypot==0.9.0
django-quill-editor==0.1.6
django-tracking-analyzer